Move EmbeddingProviderResponseValidation class to the correct package
